+++ view/fornecedor/list.php
@@ -1,18 +1,21 @@
+<?php
+ini_set('display_errors', 1);
+ini_set('display_startup_errors', 1);
+error_reporting(E_ALL);
+?>
 <div class="alert alert-warning">
     <h3>Listar Fornecedores</h3>
 </div>
 
 <div class="col-sm-12 mb-4">
-
     <div class="card shadow mb-4">
-        <!-- striped é para zebrar as linhas, cada uma com uma cor-->
         <div class="table-responsive-sm mt-4">
             <h3 class="ml-3">
                 Listar Fornecedores
                 <a class="btn btn-success float-right mb-3 mr-3" href="?p=fornecedor/add"><i class="bi bi-database-fill-add"></i></a>
             </h3>
 
-         <table class="table table-striped table-sm">
+            <table class="table table-striped table-sm">
                 <thead>
                     <tr>
                         <th>ID</th>
@@ -23,32 +26,24 @@
                 </thead>
                 <tbody>
                     <?php
-                    include_once "../Php_proj_1/models/fornecedor.php";
-                    $cat = new Fornecedor();
-                    $dados = $cat->listar(null);
+                    require_once __DIR__ . "/../../model/Fornecedor.php";
+                    require_once __DIR__ . "/../../dao/FornecedorDAO.php";
+                    
+                    $fornDAO = new FornecedorDAO();
+                    $dados = $fornDAO->listar();
+                    
                     foreach ($dados as $mostrar) {
                     ?>
                         <tr>
+                            <td><?= $mostrar["id"] ?></td>
+                            <td><?= $mostrar["nome"] ?></td>
+                            <td><?= $mostrar["cidade"] ?></td>
                             <td>
-                                <?= $mostrar["id"] ?>
-                            </td>
-                            <td>
-                                <?= $mostrar["nome"] ?>
-                            </td>
-                            <td>
-                                <?= $mostrar["cidade"] ?>
-                            </td>
-                            <td>
-                                <a
-                                    href="?p=fornecedor/edit&id=<?= $mostrar['id'] ?>"
-                                    class="btn btn-primary btn-sm">
+                                <a href="?p=fornecedor/edit&id=<?= $mostrar['id'] ?>" class="btn btn-primary btn-sm">
                                     <i class="bi bi-pencil-square"></i> Editar
                                 </a>
 
-                                <a
-                                    href="?p=fornecedor/delete&id=<?= $mostrar['id'] ?>"
-                                    class="btn btn-danger btn-sm"
-                                    onclick="return confirm('Deseja realmente excluir esta categoria?')">
+                                <a href="?p=fornecedor/delete&id=<?= $mostrar['id'] ?>" class="btn btn-danger btn-sm" onclick="return confirm('Deseja realmente excluir este fornecedor?')">
                                     <i class="bi bi-trash"></i> Excluir
                                 </a>
                             </td>
@@ -62,4 +57,4 @@
     </div>
 </div>
 
-<a href="?p=fornecedor/add" title="Add fornecedores">Add Fornecedor</a>
+<a href="?p=fornecedor/add" title="Add Fornecedor">Add Fornecedor</a>
